⚙️ [desktop-app] Add diffs and new-session slice [step 19/22]
