+++ client/desktop/lib/core/routing/desktop_router.dart
@@ -6,7 +6,9 @@ import "package:sesori_dart_core/sesori_dart_core.dart";
 
 import "../../features/auth_gate/auth_gate.dart";
 import "../../features/home/desktop_home.dart";
+import "../../features/new_session/desktop_new_session_screen.dart";
 import "../../features/projects/desktop_project_list_screen.dart";
+import "../../features/session_diffs/desktop_session_diffs_screen.dart";
 import "../../features/sessions/desktop_session_detail_screen.dart";
 import "../../features/sessions/desktop_session_list_screen.dart";
 import "../../features/settings/desktop_harnesses_settings_screen.dart";
@@ -23,121 +25,198 @@ final GlobalKey<NavigatorState> desktopRootNavigatorKey = GlobalKey<NavigatorSta
 final GoRouter desktopRouter = GoRouter(
   navigatorKey: desktopRootNavigatorKey,
   initialLocation: AppRouteDef.splash.path,
-  routes: <RouteBase>[
-    ShellRoute(
-      builder: (BuildContext context, GoRouterState state, Widget child) => AuthGate(child: child),
-      routes: <RouteBase>[
-        GoRoute(
-          path: AppRouteDef.splash.path,
-          builder: (BuildContext context, GoRouterState state) => DesktopHome(
-            onOpenProjects: () => _goRoute(context: context, route: const AppRoute.projects()),
-            onOpenSettings: () => _goRoute(context: context, route: const AppRoute.settings()),
-          ),
-        ),
-        GoRoute(
-          path: AppRouteDef.projects.path,
-          builder: (BuildContext context, GoRouterState state) => DesktopProjectListScreen(
-            onOpenSettings: () => _pushRoute(context: context, route: const AppRoute.settings()),
-            onOpenProject: ({required projectId, required projectName}) => _goRoute(
-              context: context,
-              route: AppRoute.sessions(projectId: projectId, projectName: projectName),
-            ),
-          ),
-        ),
-        GoRoute(
-          path: AppRouteDef.sessions.path,
-          builder: (BuildContext context, GoRouterState state) {
-            final route = switch (AppRoute.fromDef(
-              def: AppRouteDef.sessions,
-              pathParams: state.pathParameters,
-              queryParams: state.uri.queryParameters,
-            )) {
-              final AppRouteSessions route => route,
-              final route => throw StateError("Route ${route.def.name} is not a sessions route"),
-            };
-            return DesktopSessionListScreen(
-              projectId: route.projectId,
-              projectName: route.projectName,
-              onBack: () => _goRoute(context: context, route: const AppRoute.projects()),
-              onSessionTap: ({required session}) => _pushRoute(
-                context: context,
-                route: AppRoute.sessionDetail(
-                  projectId: route.projectId,
-                  projectName: route.projectName,
-                  sessionId: session.id,
-                  sessionTitle: session.title,
-                  readOnly: false,
-                ),
-              ),
-            );
-          },
-        ),
-        GoRoute(
-          path: AppRouteDef.sessionDetail.path,
-          builder: (BuildContext context, GoRouterState state) {
-            final route = switch (AppRoute.fromDef(
-              def: AppRouteDef.sessionDetail,
-              pathParams: state.pathParameters,
-              queryParams: state.uri.queryParameters,
-            )) {
-              final AppRouteSessionDetail route => route,
-              final route => throw StateError("Route ${route.def.name} is not a session-detail route"),
-            };
-            return DesktopSessionDetailScreen(
-              key: ValueKey((projectId: route.projectId, sessionId: route.sessionId)),
-              projectId: route.projectId,
-              sessionId: route.sessionId,
-              sessionTitle: route.sessionTitle,
-              readOnly: route.readOnly,
-              onBack: () => _popRouteOrGo(
-                context: context,
-                fallback: AppRoute.sessions(
-                  projectId: route.projectId,
-                  projectName: route.projectName,
-                ),
-              ),
-              onOpenSession: ({required projectId, required sessionId, required sessionTitle, required readOnly}) =>
-                  _pushRoute(
-                    context: context,
-                    route: AppRoute.sessionDetail(
-                      projectId: projectId,
-                      projectName: route.projectName,
-                      sessionId: sessionId,
-                      sessionTitle: sessionTitle,
-                      readOnly: readOnly,
-                    ),
-                  ),
-            );
-          },
-        ),
-        GoRoute(
-          path: AppRouteDef.settings.path,
-          builder: (BuildContext context, GoRouterState state) => DesktopSettingsScreen(
-            onClose: () => _popRoute(context: context),
-            onOpenProfile: () => _pushRoute(context: context, route: const AppRoute.settingsProfile()),
-            onOpenHarnesses: () => _pushRoute(
-              context: context,
-              route: const AppRoute.settingsHarnesses(presentation: HarnessSettingsPresentation.pushed),
-            ),
-          ),
-        ),
-        GoRoute(
-          path: AppRouteDef.settingsProfile.path,
-          builder: (BuildContext context, GoRouterState state) => DesktopProfileScreen(
-            onClose: () => _popRoute(context: context),
-            onLogoutCompleted: _goDesktopHome,
-          ),
-        ),
-        GoRoute(
-          path: AppRouteDef.settingsHarnesses.path,
-          builder: (BuildContext context, GoRouterState state) => DesktopHarnessesSettingsScreen(
-            onClose: () => _popRoute(context: context),
-          ),
-        ),
-      ],
-    ),
-  ],
+  routes: buildDesktopRoutes(),
 );
+
+@visibleForTesting
+List<RouteBase> buildDesktopRoutes() => <RouteBase>[
+  ShellRoute(
+    builder: (BuildContext context, GoRouterState state, Widget child) => AuthGate(child: child),
+    routes: <RouteBase>[
+      GoRoute(
+        path: AppRouteDef.splash.path,
+        builder: (BuildContext context, GoRouterState state) => DesktopHome(
+          onOpenProjects: () => _goRoute(context: context, route: const AppRoute.projects()),
+          onOpenSettings: () => _goRoute(context: context, route: const AppRoute.settings()),
+        ),
+      ),
+      GoRoute(
+        path: AppRouteDef.projects.path,
+        builder: (BuildContext context, GoRouterState state) => DesktopProjectListScreen(
+          onOpenSettings: () => _pushRoute(context: context, route: const AppRoute.settings()),
+          onOpenProject: ({required projectId, required projectName}) => _goRoute(
+            context: context,
+            route: AppRoute.sessions(projectId: projectId, projectName: projectName),
+          ),
+        ),
+      ),
+      GoRoute(
+        path: AppRouteDef.sessions.path,
+        builder: (BuildContext context, GoRouterState state) {
+          final route = switch (AppRoute.fromDef(
+            def: AppRouteDef.sessions,
+            pathParams: state.pathParameters,
+            queryParams: state.uri.queryParameters,
+          )) {
+            final AppRouteSessions route => route,
+            final route => throw StateError("Route ${route.def.name} is not a sessions route"),
+          };
+          return DesktopSessionListScreen(
+            projectId: route.projectId,
+            projectName: route.projectName,
+            onBack: () => _goRoute(context: context, route: const AppRoute.projects()),
+            onSessionTap: ({required session}) => _pushRoute(
+              context: context,
+              route: AppRoute.sessionDetail(
+                projectId: route.projectId,
+                projectName: route.projectName,
+                sessionId: session.id,
+                sessionTitle: session.title,
+                readOnly: false,
+              ),
+            ),
+            onNewSession: () => _pushRoute(
+              context: context,
+              route: AppRoute.newSession(projectId: route.projectId, projectName: route.projectName),
+            ),
+          );
+        },
+      ),
+      GoRoute(
+        path: AppRouteDef.newSession.path,
+        builder: (BuildContext context, GoRouterState state) {
+          final route = switch (AppRoute.fromDef(
+            def: AppRouteDef.newSession,
+            pathParams: state.pathParameters,
+            queryParams: state.uri.queryParameters,
+          )) {
+            final AppRouteNewSession route => route,
+            final route => throw StateError("Route ${route.def.name} is not a new-session route"),
+          };
+          return DesktopNewSessionScreen(
+            projectId: route.projectId,
+            projectName: route.projectName,
+            onBack: () => _popRouteOrGo(
+              context: context,
+              fallback: AppRoute.sessions(projectId: route.projectId, projectName: route.projectName),
+            ),
+            onOpenHarnessSettings: () => _pushRoute(
+              context: context,
+              route: const AppRoute.settingsHarnesses(presentation: HarnessSettingsPresentation.modal),
+            ),
+            onSessionCreated: ({required session}) => _replaceRoute(
+              context: context,
+              route: AppRoute.sessionDetail(
+                projectId: route.projectId,
+                projectName: route.projectName,
+                sessionId: session.id,
+                sessionTitle: session.title,
+                readOnly: false,
+              ),
+            ),
+          );
+        },
+      ),
+      GoRoute(
+        path: AppRouteDef.sessionDetail.path,
+        builder: (BuildContext context, GoRouterState state) {
+          final route = switch (AppRoute.fromDef(
+            def: AppRouteDef.sessionDetail,
+            pathParams: state.pathParameters,
+            queryParams: state.uri.queryParameters,
+          )) {
+            final AppRouteSessionDetail route => route,
+            final route => throw StateError("Route ${route.def.name} is not a session-detail route"),
+          };
+          return DesktopSessionDetailScreen(
+            key: ValueKey((projectId: route.projectId, sessionId: route.sessionId)),
+            projectId: route.projectId,
+            sessionId: route.sessionId,
+            sessionTitle: route.sessionTitle,
+            readOnly: route.readOnly,
+            onBack: () => _popRouteOrGo(
+              context: context,
+              fallback: AppRoute.sessions(
+                projectId: route.projectId,
+                projectName: route.projectName,
+              ),
+            ),
+            onShowDiffs: () => _pushRoute(
+              context: context,
+              route: AppRoute.sessionDiffs(
+                projectId: route.projectId,
+                projectName: route.projectName,
+                sessionId: route.sessionId,
+              ),
+            ),
+            onOpenSession: ({required projectId, required sessionId, required sessionTitle, required readOnly}) =>
+                _pushRoute(
+                  context: context,
+                  route: AppRoute.sessionDetail(
+                    projectId: projectId,
+                    projectName: route.projectName,
+                    sessionId: sessionId,
+                    sessionTitle: sessionTitle,
+                    readOnly: readOnly,
+                  ),
+                ),
+          );
+        },
+      ),
+      GoRoute(
+        path: AppRouteDef.sessionDiffs.path,
+        builder: (BuildContext context, GoRouterState state) {
+          final route = switch (AppRoute.fromDef(
+            def: AppRouteDef.sessionDiffs,
+            pathParams: state.pathParameters,
+            queryParams: state.uri.queryParameters,
+          )) {
+            final AppRouteSessionDiffs route => route,
+            final route => throw StateError("Route ${route.def.name} is not a session-diffs route"),
+          };
+          return DesktopSessionDiffsScreen(
+            projectId: route.projectId,
+            sessionId: route.sessionId,
+            onBack: () => _popRouteOrGo(
+              context: context,
+              fallback: AppRoute.sessionDetail(
+                projectId: route.projectId,
+                projectName: route.projectName,
+                sessionId: route.sessionId,
+                sessionTitle: null,
+                readOnly: false,
+              ),
+            ),
+          );
+        },
+      ),
+      GoRoute(
+        path: AppRouteDef.settings.path,
+        builder: (BuildContext context, GoRouterState state) => DesktopSettingsScreen(
+          onClose: () => _popRoute(context: context),
+          onOpenProfile: () => _pushRoute(context: context, route: const AppRoute.settingsProfile()),
+          onOpenHarnesses: () => _pushRoute(
+            context: context,
+            route: const AppRoute.settingsHarnesses(presentation: HarnessSettingsPresentation.pushed),
+          ),
+        ),
+      ),
+      GoRoute(
+        path: AppRouteDef.settingsProfile.path,
+        builder: (BuildContext context, GoRouterState state) => DesktopProfileScreen(
+          onClose: () => _popRoute(context: context),
+          onLogoutCompleted: _goDesktopHome,
+        ),
+      ),
+      GoRoute(
+        path: AppRouteDef.settingsHarnesses.path,
+        builder: (BuildContext context, GoRouterState state) => DesktopHarnessesSettingsScreen(
+          onClose: () => _popRoute(context: context),
+        ),
+      ),
+    ],
+  ),
+];
 
 void _goDesktopHome() {
   // ignore: no_slop_linter/avoid_raw_go_router, desktop router's typed route boundary
@@ -152,6 +231,11 @@ void _goRoute({required BuildContext context, required AppRoute route}) {
 void _pushRoute({required BuildContext context, required AppRoute route}) {
   // ignore: no_slop_linter/avoid_raw_go_router, desktop router's typed route boundary
   unawaited(GoRouter.of(context).push<void>(route.buildPath()));
+}
+
+void _replaceRoute({required BuildContext context, required AppRoute route}) {
+  // ignore: no_slop_linter/avoid_raw_go_router, desktop router's typed route boundary
+  GoRouter.of(context).replace<void>(route.buildPath());
 }
 
 void _popRouteOrGo({required BuildContext context, required AppRoute fallback}) {
