+++ client/module_app_ui/lib/sesori_app_ui.dart
@@ -6,6 +6,8 @@ export "src/extensions/build_context_x.dart";
 export "src/extensions/remote_failure_x.dart";
 export "src/extensions/status_colors.dart";
 export "src/extensions/text_style_x.dart";
+export "src/features/new_session/new_session_plugin_chooser.dart";
+export "src/features/new_session/new_session_view.dart";
 export "src/features/project_list/add_project_dialog.dart";
 export "src/features/project_list/new_folder_dialog.dart";
 export "src/features/project_list/project_list_view.dart";
@@ -51,6 +53,7 @@ export "src/features/session_detail/widgets/text_part_widget.dart";
 export "src/features/session_detail/widgets/tool_part_widget.dart";
 export "src/features/session_detail/widgets/user_message_card.dart";
 export "src/features/session_detail/widgets/voice_cancel_button.dart";
+export "src/features/session_diffs/session_diffs_view.dart";
 export "src/features/session_list/pr_status_row.dart";
 export "src/features/session_list/rename_session_dialog.dart";
 export "src/features/session_list/session_archived_empty_state.dart";
