+++ client/module_app_ui/lib/src/features/new_session/new_session_no_harness_notice.dart
@@ -1,6 +1,7 @@
 import "package:material_ui/material_ui.dart";
-import "package:sesori_app_ui/sesori_app_ui.dart";
 import "package:theme_prego/module_prego.dart";
+
+import "../../extensions/build_context_x.dart";
 
 /// What the options block says when the bridge answered with no harness at all.
 ///
