+++ client/module_app_ui/lib/src/features/new_session/new_session_plugin_chooser.dart
@@ -1,8 +1,9 @@
 import "package:material_ui/material_ui.dart";
-import "package:sesori_app_ui/sesori_app_ui.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:theme_prego/module_prego.dart";
+
+import "../../extensions/build_context_x.dart";
 
 /// The harness the new session will run on, shown as the picked harness' mark
 /// and name over an anchored menu listing every harness this bridge reports.
