+++ client/module_app_ui/lib/src/features/session_diffs/models/diff_file_view_model.dart
