+++ client/module_app_ui/lib/src/features/session_diffs/models/diff_view_model_builder.dart
