+++ client/module_app_ui/lib/src/features/session_diffs/utils/diff_theme.dart
