+++ client/module_app_ui/lib/src/features/session_diffs/widgets/diff_file_header_delegate.dart
