+++ client/module_app_ui/test/features/session_diffs/widgets/diff_file_widget_test.dart
@@ -1,7 +1,7 @@
 import "package:flutter_test/flutter_test.dart";
 import "package:material_ui/material_ui.dart";
-import "package:sesori_mobile/features/session_diffs/models/diff_file_view_model.dart";
-import "package:sesori_mobile/features/session_diffs/widgets/diff_file_widget.dart";
+import "package:sesori_app_ui/src/features/session_diffs/models/diff_file_view_model.dart";
+import "package:sesori_app_ui/src/features/session_diffs/widgets/diff_file_widget.dart";
 import "package:sesori_shared/sesori_shared.dart";
 
 void main() {
